pkg/rundeck/responses/toggle_test.go
package responses

import (
	"encoding/json"
	"testing"

	"github.com/lusis/go-rundeck/pkg/rundeck/responses/testdata"
	"github.com/mitchellh/mapstructure"
	"github.com/stretchr/testify/assert"
)

func TestBulkToggleResponse(t *testing.T) {
	obj := &BulkToggleResponse{}
	data, dataErr := testdata.GetBytes(BulkToggleResponseTestFile)
	if dataErr != nil {
		t.Fatalf(dataErr.Error())
	}
	placeholder := make(map[string]interface{})
	_ = json.Unmarshal(data, &placeholder)
	config := newMSDecoderConfig()
	config.Result = obj
	decoder, newErr := mapstructure.NewDecoder(config)
	assert.NoError(t, newErr)
	dErr := decoder.Decode(placeholder)
	assert.NoError(t, dErr)
	assert.Implements(t, (*VersionedResponse)(nil), obj)
}
